Update RUN commands of unmanaged tests.
This was partially scripted, but there are enough special cases that it
deserves some manual review.

typedefs, typedefbounds: Turn the commented-out commands for an
idempotence test into a real test, since it passes.

Addition of "rm -rf %t*" lines is split into a separate commit to make
this diff easier for naive diff tools (e.g., GitHub) to review by
avoiding breaking the alignment of corresponding old and new commands.