Fix error in canBeNtArray making ensureNtCorrect add incorrect constraints
This function checked if a type was a PointerType or ArrayType, but it
could be a wrapper type around one of these types. For example, ParenType
and Decayed type instances were both treated as not pointers or arrays
even if the type they wrapped was a pointer or array type.