Fix for empty _Nt_checked
Checked C doesn't allow _Nt_checked[0], so add a constraint that force
constant size arrays with size 0 to be regular arrays. This resovles a
bug that existed prior to this PR, and avoids a integer underflow
in that could happen when subtracting 1 from 0 while adjusting the
bounds of a _Nt_checked.