Remove the mkString fixes from this PR.
They are no longer important to have in this PR now that I have switched
unchanged multi-decl members back from mkString to Decl::print in most
cases. I can submit the fixes in a separate PR if we like, or we can
hold off and maybe we'll do a broader overhaul of mkString.